case_ai_engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "case-ai-engine"
version = "0.1.0"
description = "Intelligent CASE module for FastAPI project generation from JSON/YAML spec"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "typer>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.4",
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["engine"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["engine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
